The first Mindcraft hotfix!! \nn/
The Issue is the Url. It is right to use a "http://bla.com" or a "http://www.bla.com" format, as it is the same. However it IS important to use the same construction everywhere.
Otherwise, some browsers (Ahhhchromehhhem) might consider it a reason not to work. Lazy browsers they are, aren't they?